requirements.txt
flask>=3.0.0
gunicorn
psycopg2-binary
python-dotenv
